Correcciones finales

+++ PaginaWeb/admin/utilidades/modificar-carta-check.php
@@ -13,6 +13,16 @@ function no_hay_carta() {
     alert("Primero debe crear y seleccionar una carta para modificar");
     window.location.replace("../dashboard.php");
 }
+function ya_existe() {
+    alert("Ya existe una carta con estos mismos datos");
+    // window.location.replace("ingresar-carta.php");
+    history.back();
+}
+function caracteres_ilegales() {
+    alert("No se pueden utilizar ciertos símbolos ingresados");
+    // window.location.replace("ingresar-set.php");
+    history.back();
+}
 </script>
 <?php
     session_start();
@@ -20,9 +30,45 @@ function no_hay_carta() {
     if (! empty($_POST)) {
         if (isset($_POST['carta'])) {
             if (isset($_POST['nuevodato']) && strlen(trim($_POST['nuevodato'])) > 0){
-                $carta = explode('?',$_POST['carta']);
-                pg_exec("select modificar_atributos_carta('".$_POST['nuevodato']."','".$_POST['atributo']."','".$carta[0]."','".$carta[1]."','".$carta[2]."')") or die('Consulta fallida');
-                echo "<script>correcto();</script>";
+                if (preg_match('/[#$%^*()+=\\[\]\';,.\/{}|":<>?~\\\\]/', $_POST['nuevodato']) == 0 or $_POST['atributo'] == 'imagen') {
+                        switch ($_POST['atributo']) {
+                            case 'estampado':
+                                $carta = explode('?',$_POST['carta']);
+                                $existe_la_carta = pg_exec("select * from buscar_carta_en_sistema('".$carta[0]."','".$_POST['nuevodato']."','".$carta[1]."')") or die('Consulta fallida');
+                                if (pg_fetch_result($existe_la_carta,'buscar_carta_en_sistema') == '0') {
+                                    pg_exec("select modificar_atributos_carta('".$_POST['nuevodato']."','".$_POST['atributo']."','".$carta[0]."','".$carta[1]."','".$carta[2]."')") or die('Consulta fallida');
+                                    echo "<script>correcto();</script>";
+                                } else {
+                                    echo "<script>ya_existe();</script>";
+                                }
+
+                            case 'idioma':
+                                $carta = explode('?',$_POST['carta']);
+                                $existe_la_carta = pg_exec("select * from buscar_carta_en_sistema('".$carta[0]."','".$carta[2]."','".$_POST['nuevodato']."')") or die('Consulta fallida');
+                                if (pg_fetch_result($existe_la_carta,'buscar_carta_en_sistema') == '0') {
+                                    pg_exec("select modificar_atributos_carta('".$_POST['nuevodato']."','".$_POST['atributo']."','".$carta[0]."','".$carta[1]."','".$carta[2]."')") or die('Consulta fallida');
+                                    echo "<script>correcto();</script>";
+                                } else {
+                                    echo "<script>ya_existe();</script>";
+                                }
+                            case 'id':
+                                $carta = explode('?',$_POST['carta']);
+                                $existe_la_carta = pg_exec("select * from buscar_carta_en_sistema('".$_POST['nuevodato']."','".$carta[2]."','".$carta[1]."')") or die('Consulta fallida');
+                                if (pg_fetch_result($existe_la_carta,'buscar_carta_en_sistema') == '0') {
+                                    pg_exec("select modificar_atributos_carta('".$_POST['nuevodato']."','".$_POST['atributo']."','".$carta[0]."','".$carta[1]."','".$carta[2]."')") or die('Consulta fallida');
+                                    echo "<script>correcto();</script>";
+                                } else {
+                                    echo "<script>ya_existe();</script>";
+                                }
+                            default:
+                                $carta = explode('?',$_POST['carta']);
+                                pg_exec("select modificar_atributos_carta('".$_POST['nuevodato']."','".$_POST['atributo']."','".$carta[0]."','".$carta[1]."','".$carta[2]."')") or die('Consulta fallida');
+                                echo "<script>correcto();</script>";
+                        }
+                } else {
+                    echo "<script>caracteres_ilegales();</script>";
+                }
+
             } else {
                 echo "<script>espacios_vacios();</script>";
             }
